requirements.txt
# Core Agent Framework
google-adk>=1.19.0
litellm>=1.80.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Data & Database
snowflake-snowpark-python>=1.11.0
google-cloud-firestore>=2.14.0
google-cloud-secret-manager>=2.17.0

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-instrumentation-fastapi>=0.43b0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
httpx>=0.26.0
tenacity>=8.2.3
pybreaker>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0

# Logging
structlog>=24.1.0
python-json-logger>=2.0.7